+++ cpp/grid_view.cpp
@@ -12,10 +12,10 @@ void implement_grid_view(jlcxx::Module& module)
     .method("push_front!", [](GridView& view, void* widget) -> void {
         view.push_front(*((Widget*) widget));
     })
-    .method("insert!", [](GridView& view, size_t i, void* widget) -> void {
+    .method("insert!", [](GridView& view, uint64_t i, void* widget) -> void {
         view.insert(*((Widget*) widget), i);
     })
-    .method("remove!", [](GridView& view, size_t index) -> void {
+    .method("remove!", [](GridView& view, uint64_t index) -> void {
         view.remove(index);
     })
     .method("clear!", [](GridView& view) -> void {
